Updated deploy.bash  - added a check for the celery environment

+++ scripts/deploy.bash
@@ -79,13 +79,17 @@ fi
 
 # let gunicorn know its ok to restart
 if ! [ "$JOBS_NODE" ]; then
-    echo "- gunicorn"
-    for pid in $(pgrep -fl "gunicorn: worke" | awk '{print $1}'); do
-    sudo kill -1 $pid
-    sleep 0.5
-    done
-    echo "- celery"
-    sudo systemctl celery restart
+    if ! [[ -z CELERY_NODE  ]];
+    then
+      echo "- gunicorn"
+      for pid in $(pgrep -fl "gunicorn: worke" | awk '{print $1}'); do
+      sudo kill -1 $pid
+      sleep 0.5
+      done
+    else
+      echo "- celery"
+      sudo systemctl restart celery
+    fi
 fi
 
 # invalidate cloudfront
